- separated parallel_for_internals.hpp and parallel_for.hpp - added the possibility to control the grain size in the static scheduling   policy of the ParallelFor(Reduce) pattern(s) - ff_node::eosnotify interface changed:    virtual void eosnotify(int id=-1) --> virtual void eosnotify(ssize_t id=-1)

+++ tests/pipe_basic.cpp
@@ -168,7 +168,7 @@ int main() {
             }
             return GO_ON;
         }            
-        void eosnotify(int id) {
+        void eosnotify(ssize_t id) {
             if (id==-1) lb->broadcast_task(EOS);
         }
     };
